+++ CHANGELOG.md
@@ -9,8 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+## [0.3.4] - 2022-04-10
+
+### Added
+
 - Support `@description` decorator for output elements
 - Support public module registry references
+- Support module alias references
 
 ## [0.3.3] - 2022-03-21
 
@@ -88,7 +93,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release to be able to parse all the official examples of Bicep [101](https://github.com/Azure/bicep/tree/main/docs/examples/101).
 
-[Unreleased]: https://github.com/gruebel/pycep/compare/0.3.3...HEAD
+[Unreleased]: https://github.com/gruebel/pycep/compare/0.3.4...HEAD
+[0.3.4]: https://github.com/gruebel/pycep/compare/0.3.3...0.3.4
 [0.3.3]: https://github.com/gruebel/pycep/compare/0.3.2...0.3.3
 [0.3.2]: https://github.com/gruebel/pycep/compare/0.3.1...0.3.2
 [0.3.1]: https://github.com/gruebel/pycep/compare/0.3.0...0.3.1